Added Breakfast salt counr
I changed the prefabs to include the buttons so upon the button press the health bar changes. Each works but I plan to merge the change panels script to the Item base script so things get done in one script. Or I'll change it so the change panel script can be called from anywhere